Move CPlayState traits
